Comment on tests?
I’m not sure I’m writing the tests correctly. I wrote the first couple
with my tutor.

I think I’m still struggling with the concept of testing. I could use
some help figuring out what to test for and then once I do, with any
luck, I can figure out the right code to do it.

I’m not trying to do any validation on the user input right? This is
strictly testing my own code for weak points. For example, for
getDayName, would I want to create a var for a string like “Thursday”
and then compare it to a date in a format like “May 14 2015” if that
day is a Thursday?