final commit
Here are my final changes after watching your video. It’s good to know
I was making it way too complicated. I think I have a better feel for
what to do now. I have one question about testing isFuture. If the date
is today, it’s also not in the future. Do we not test for that in
isFuture because we’re already testing for it isToday?